+++ Sources/SuperwallKit/Analytics/Superwall Placement/SuperwallEvent.swift
@@ -207,6 +207,15 @@ public enum SuperwallEvent {
   /// When the shimmer view stops showing.
   case shimmerViewComplete
 
+  /// When the redemption of a code is initiated.
+  case redemptionStart
+
+  /// When the redemption of a code completes.
+  case redemptionComplete
+
+  /// When the redemption of a code fails.
+  case redemptionFail
+
   /// When the enrichment request starts.
   case enrichmentStart
 
@@ -367,6 +376,12 @@ extension SuperwallEvent {
       return .init(objcEvent: .shimmerViewStart)
     case .shimmerViewComplete:
       return .init(objcEvent: .shimmerViewComplete)
+    case .redemptionStart:
+      return .init(objcEvent: .redemptionStart)
+    case .redemptionComplete:
+      return .init(objcEvent: .redemptionComplete)
+    case .redemptionFail:
+      return .init(objcEvent: .redemptionFail)
     case .enrichmentFail:
       return .init(objcEvent: .enrichmentFail)
     case .enrichmentStart: