+++ Sources/SuperwallKit/Network/API.swift
@@ -12,12 +12,20 @@ enum EndpointHost {
   case collector
   case enrichment
   case adServices
+  case web2app
 }
 
 protocol ApiHostConfig {
+  var networkEnvironment: SuperwallOptions.NetworkEnvironment { get }
   var port: Int? { get }
   var scheme: String { get }
   var host: String { get }
+  var path: String { get }
+}
+
+extension ApiHostConfig {
+  var port: Int? { return networkEnvironment.port }
+  var scheme: String { return networkEnvironment.scheme }
 }
 
 struct Api {
@@ -25,13 +33,14 @@ struct Api {
   let collector: Collector
   let enrichment: Enrichment
   let adServices: AdServices
-  static let version1 = "/api/v1/"
+  let web2app: Web2App
 
   init(networkEnvironment: SuperwallOptions.NetworkEnvironment) {
     base = Base(networkEnvironment: networkEnvironment)
     collector = Collector(networkEnvironment: networkEnvironment)
     enrichment = Enrichment(networkEnvironment: networkEnvironment)
     adServices = AdServices(networkEnvironment: networkEnvironment)
+    web2app = Web2App(networkEnvironment: networkEnvironment)
   }
 
   func getConfig(host: EndpointHost) -> ApiHostConfig {
@@ -44,14 +53,15 @@ struct Api {
       return enrichment
     case .adServices:
       return adServices
+    case .web2app:
+      return web2app
     }
   }
 
   struct Base: ApiHostConfig {
-    private let networkEnvironment: SuperwallOptions.NetworkEnvironment
-    var port: Int? { return networkEnvironment.port }
-    var scheme: String { return networkEnvironment.scheme }
+    let networkEnvironment: SuperwallOptions.NetworkEnvironment
     var host: String { return networkEnvironment.baseHost }
+    var path: String { return "/api/v1/" }
 
     init(networkEnvironment: SuperwallOptions.NetworkEnvironment) {
       self.networkEnvironment = networkEnvironment
@@ -59,21 +69,23 @@ struct Api {
   }
 
   struct Collector: ApiHostConfig {
-    private let networkEnvironment: SuperwallOptions.NetworkEnvironment
-    var port: Int? { return networkEnvironment.port }
-    var scheme: String { return networkEnvironment.scheme }
+    let networkEnvironment: SuperwallOptions.NetworkEnvironment
     var host: String { return networkEnvironment.collectorHost }
+    var path: String { return "/api/v1/" }
 
     init(networkEnvironment: SuperwallOptions.NetworkEnvironment) {
       self.networkEnvironment = networkEnvironment
     }
   }
 
+
+
   struct Enrichment: ApiHostConfig {
-    private let networkEnvironment: SuperwallOptions.NetworkEnvironment
+    internal let networkEnvironment: SuperwallOptions.NetworkEnvironment
     var port: Int? { return networkEnvironment.port }
     var scheme: String { return networkEnvironment.scheme }
     var host: String { return networkEnvironment.enrichmentHost }
+    var path: String { return "/api/v1/" }
 
     init(networkEnvironment: SuperwallOptions.NetworkEnvironment) {
       self.networkEnvironment = networkEnvironment
@@ -81,10 +93,19 @@ struct Api {
   }
 
   struct AdServices: ApiHostConfig {
-    private let networkEnvironment: SuperwallOptions.NetworkEnvironment
-    var port: Int? { return networkEnvironment.port }
-    var scheme: String { return networkEnvironment.scheme }
+    let networkEnvironment: SuperwallOptions.NetworkEnvironment
     var host: String { return networkEnvironment.adServicesHost }
+    var path: String { return "/api/v1/" }
+
+    init(networkEnvironment: SuperwallOptions.NetworkEnvironment) {
+      self.networkEnvironment = networkEnvironment
+    }
+  }
+
+  struct Web2App: ApiHostConfig {
+    let networkEnvironment: SuperwallOptions.NetworkEnvironment
+    var host: String { return networkEnvironment.web2AppHost }
+    var path: String { return "/subscriptions-api/public/v1/" }
 
     init(networkEnvironment: SuperwallOptions.NetworkEnvironment) {
       self.networkEnvironment = networkEnvironment