+++ Sources/SuperwallKit/Network/Custom URL Session/CustomURLSession.swift
@@ -90,6 +90,7 @@ class CustomURLSession {
         "request_duration": requestDuration
       ]
     )
+
     guard let value = try? Kind.jsonDecoder.decode(
       Response.self,
       from: data